requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pydantic-settings
openai
chromadb
PyMuPDF
langchain
langchain-community
tiktoken
python-multipart
